Move the Intro Card into its own component to resolve reac hook issue "Error: React Hook "useTransform" cannot be called inside a callback. React Hooks must be called in a React function component or a custom React Hook function.
